refactoring (ai): move errors to src/error
