+++ packages/ai/core/generate-text/output.test.ts
@@ -1,6 +1,6 @@
 import { fail } from 'assert';
 import { z } from 'zod';
-import { verifyNoObjectGeneratedError } from '../../errors/no-object-generated-error';
+import { verifyNoObjectGeneratedError } from '../../src/error/no-object-generated-error';
 import { object } from './output';
 import { FinishReason } from '../types';
