+++ packages/ai/core/generate-text/parse-tool-call.ts
@@ -1,8 +1,8 @@
 import { LanguageModelV2ToolCall } from '@ai-sdk/provider';
 import { safeParseJSON, safeValidateTypes } from '@ai-sdk/provider-utils';
-import { InvalidToolArgumentsError } from '../../errors/invalid-tool-arguments-error';
-import { NoSuchToolError } from '../../errors/no-such-tool-error';
-import { ToolCallRepairError } from '../../errors/tool-call-repair-error';
+import { InvalidToolArgumentsError } from '../../src/error/invalid-tool-arguments-error';
+import { NoSuchToolError } from '../../src/error/no-such-tool-error';
+import { ToolCallRepairError } from '../../src/error/tool-call-repair-error';
 import { ModelMessage } from '../prompt';
 import { asSchema } from '../util';
 import { ToolCallUnion } from './tool-call';
