+++ packages/ai/core/generate-text/run-tools-transformation.test.ts
@@ -5,7 +5,7 @@ import {
   convertReadableStreamToArray,
 } from '@ai-sdk/provider-utils/test';
 import { z } from 'zod';
-import { NoSuchToolError } from '../../errors';
+import { NoSuchToolError } from '../../src/error/no-such-tool-error';
 import { MockTracer } from '../test/mock-tracer';
 import { runToolsTransformation } from './run-tools-transformation';
