+++ packages/ai/core/prompt/prepare-call-settings.ts
@@ -1,4 +1,4 @@
-import { InvalidArgumentError } from '../../errors/invalid-argument-error';
+import { InvalidArgumentError } from '../../src/error/invalid-argument-error';
 import { CallSettings } from './call-settings';
 
 /**
