+++ packages/ai/core/tool/mcp/mcp-sse-transport.ts
@@ -1,5 +1,5 @@
 import { createEventSourceParserStream } from '@ai-sdk/provider-utils';
-import { MCPClientError } from '../../../errors';
+import { MCPClientError } from '../../../src/error/mcp-client-error';
 import { JSONRPCMessage, JSONRPCMessageSchema } from './json-rpc-message';
 import { MCPTransport } from './mcp-transport';
