+++ packages/ai/core/util/cosine-similarity.ts
@@ -1,4 +1,4 @@
-import { InvalidArgumentError } from '../../errors/invalid-argument-error';
+import { InvalidArgumentError } from '../../src/error/invalid-argument-error';
 
 /**
  * Calculates the cosine similarity between two vectors. This is a useful metric for
