+++ packages/ai/mcp-stdio/mcp-stdio-transport.ts
@@ -5,7 +5,7 @@ import {
   JSONRPCMessageSchema,
 } from '../core/tool/mcp/json-rpc-message';
 import { MCPTransport } from '../core/tool/mcp/mcp-transport';
-import { MCPClientError } from '../errors';
+import { MCPClientError } from '../src/error/mcp-client-error';
 import { createChildProcess } from './create-child-process';
 
 export interface StdioConfig {
