+++ packages/ai/src/error/invalid-argument-error.ts
