+++ packages/ai/src/error/invalid-tool-arguments-error.ts
