+++ packages/ai/src/error/no-output-specified-error.ts
